requirements.txt
flask
langchain
openai
gunicorn
streamlit
requests
python-dotenv
python-math
pandas
TIME-python
spacy
vectordb
wikipedia
transformers==4.35.2
torch
docarray
tqdm
nltk
fastapi
uvicorn[standard]
tensorflow[and-cuda]
accelerate
bitsandbytes
scipy
google-search-results
pymongo
openrouteservice
sentencepiece
colorlog
groq
langgraph
langchain_community
langchain-groq
tavily-python
langgraph-checkpoint-sqlite
geopy